+++ src/main/java/pers/goetboy/entity/sys/RoleUser.java
@@ -18,26 +18,26 @@ public class RoleUser extends AbstractEntity implements Serializable {
      * 角色id
      */
     @Column(name = "role_id")
-    private Integer roleId;
+    private Long roleId;
     /**
      * 用户id
      */
     @Column(name = "user_id")
-    private Integer userId;
+    private Long userId;
 
-    public Integer getRoleId() {
+    public Long getRoleId() {
         return roleId;
     }
 
-    public void setRoleId(Integer roleId) {
+    public void setRoleId(Long roleId) {
         this.roleId = roleId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
